error pages mostly done
Update opions.ts

Update errorPage.ts

Update main.ts
